Run interop tests on CI (#19)
This PR adds the beginning of an implementation of the gRPC interoperability tests, and a first pass at running them on CI. Although this implementation is far from complete, I want to get the tests that _are_ currently implemented merged in so we can start running them against other changes.

## Current state of test implementation:

We have a test client implementation with a framework for adding more tests. We have not yet implemented a test server, and the `docker-compose` file for running the tests currently runs them against `grpc-go`'s test server for now.

### Test checklist

- [x] `empty_unary`: implemented in client
- [ ] `cacheable_unary`: started, requires request context implementation to set cacheable flag
- [x] `large_unary`: implemented in client, broken due to #14
- [ ] ~`client_compressed_unary`~: requires gRPC compression, NYI
- [ ] ~`server_compressed_unary`~: requires gRPC compression, NYI
- [x] `client_streaming`: implemented in client, broken due to #16
- [ ] ~`client_compressed_streaming`~: requires gRPC compression, NYI
- [ ] `server_streaming`
- [ ] ~`server_compressed_streaming`~: requires gRPC compression, NYI
- [ ] `ping_pong`
- [ ] `empty_stream`
- [ ] ~`compute_engine_creds`~ requires auth, NYI
- [ ] ~`jwt_token_creds`~ requires auth, NYI
- [ ] ~`oauth2_auth_token`~ requires auth, NYI
- [ ] ~`per_rpc_creds`~ requires auth, NYI
- [ ] `custom_metadata`
- [ ] `status_code_and_message`
- [ ] `unimplemented_method`
- [ ] `unimplemented_service`
- [ ] `cancel_after_begin`
- [ ] `cancel_after_first_response`
- [ ] `timeout_on_sleeping_server`
- [ ] ~`concurrent_large_unary`~: not yet added to test case descriptions

I've added a `docker-compose.yml` and two `Dockerfile`s for running the test client against the `grpc-go` test server without requiring a Go installation on the host machine. This is now used in the `.travis.yml` for running the tests, and might be useful in dev as well. I also added a quick README briefly documenting how to run the interop tests.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>