Add support for binary custom metadata fields (#102)
Introduce _bin APIs for dealing with binary metadata keys, and
distinct Binary vs Ascii MetadataKey types. Binary values are
base64 encoded.

With the split of binary vs ascii keys I can't think of a reasonable way
to implement this that doesn't have an API that is so complicated that
it's not worth using for the very small performance benefits it provides.
It's still possible to go down to the lower HeaderMap level and use drain
from there.